feat: added lint-staged, and updated eslint config to airbnb standard
