+++ .lintstagedrc
@@ -1,0 +1,3 @@
+{
+  "*.{js,ts,jsx,tsx}": ["npm run lint --silent"]
+}
