+++ package.json
@@ -9,16 +9,13 @@
   "devDependencies": {
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
-    "@types/eslint": "^7.2.6",
-    "@types/eslint-plugin-prettier": "^3.1.0",
-    "@types/prettier": "^2.1.6",
     "@types/rimraf": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.14.1",
-    "@typescript-eslint/parser": "^4.14.1",
-    "eslint": "^7.18.0",
+    "@typescript-eslint/eslint-plugin": "^4.16.1",
+    "@typescript-eslint/parser": "^4.16.1",
+    "eslint": "^7.21.0",
     "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-airbnb-typescript": "^12.0.0",
-    "eslint-config-prettier": "^7.2.0",
+    "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.3.1",
@@ -31,7 +28,8 @@
     "rollup": "^2.38.1",
     "rollup-plugin-typescript2": "^0.29.0",
     "ts-jest": "^26.5.0",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "lint-staged": "^10.5.4"
   },
   "files": [
     "dist"
@@ -52,11 +50,13 @@
   },
   "scripts": {
     "build": "rimraf dist && rollup -c",
-    "lint": "eslint . --ext .ts --cache --fix",
+    "lint": "eslint . --fix",
     "start": "node dist/index.js",
     "test": "jest",
     "test:watch": "jest --watchAll",
     "watch": "rollup -cw"
   },
-  "types": "dist/index.d.ts"
+  "types": "dist/index.d.ts",
+  "dependencies": {
+  }
 }
